Add tooltip text in settings menu

+++ 3dEngine/src/scene/ui/widget/text/Text.cpp
@@ -32,7 +32,7 @@ namespace urchin {
     }
 
     Text::~Text() {
-        if (isTranslatableLabel()) {
+        if (isTranslatableLabel() && i18nService) {
             i18nService->remove(this);
         }
     }
